refactor: move catalog contracts to session crate
Move catalog, schema, table provider, and table function contracts into datafusion-session while preserving datafusion-catalog re-exports. Expose catalog access as a required Session capability, including across ForeignSession and the FFI boundary.

AI Disclosure: This code was written in part by an AI agent.
