+++ datafusion/ffi/src/tests/async_provider.rs
@@ -31,7 +31,7 @@ use std::sync::Arc;
 use arrow::array::RecordBatch;
 use arrow::datatypes::Schema;
 use async_trait::async_trait;
-use datafusion_catalog::TableProvider;
+use datafusion_catalog::{MemoryCatalogProvider, TableProvider};
 use datafusion_common::{Result, exec_err};
 use datafusion_execution::RecordBatchStream;
 use datafusion_expr::Expr;
@@ -135,11 +135,28 @@ impl TableProvider for AsyncTableProvider {
 
     async fn scan(
         &self,
-        _state: &dyn Session,
+        state: &dyn Session,
         _projection: Option<&Vec<usize>>,
         _filters: &[Expr],
         _limit: Option<usize>,
     ) -> Result<Arc<dyn ExecutionPlan>> {
+        let catalog = state.catalog_list().catalog("datafusion").ok_or_else(|| {
+            datafusion_common::exec_datafusion_err!("missing datafusion catalog")
+        })?;
+        let schema = catalog.schema("public").ok_or_else(|| {
+            datafusion_common::exec_datafusion_err!("missing public schema")
+        })?;
+        if schema.table("external_table").await?.is_none() {
+            return exec_err!("missing external_table");
+        }
+
+        // Register a catalog from the dynamically loaded library so the host
+        // can verify that catalog mutations cross the FFI boundary as well.
+        state.catalog_list().register_catalog(
+            "ffi_registered".to_owned(),
+            Arc::new(MemoryCatalogProvider::new()),
+        );
+
         Ok(Arc::new(AsyncTestExecutionPlan::new(
             self.batch_request.clone(),
             self.batch_receiver.resubscribe(),
